+++ source/toolkits/TranslatorTk.h
@@ -22,6 +22,7 @@ class TranslatorTk
 		static std::string stringVecToString(const StringVec& vec, std::string separator);
 		static unsigned fadviseArgsStrToFlags(std::string fadviseArgsStr);
 		static unsigned madviseArgsStrToFlags(std::string madviseArgsStr);
+		static std::string intVectoHumanStr(const IntVec& intVec);
 };